+++ CHANGELOG.md
@@ -40,6 +40,10 @@ Major release: Python 3.12+ modernization
 
 ### Added
 
+- Add the `deprecated.google` module: the `@deprecated`, `@versionadded` and `@versionchanged`
+  decorators insert a `Deprecated:`, `Version added:` or `Version changed:` section
+  in [Google style](https://google.github.io/styleguide/pyguide.html#38-comments-and-docstrings)
+  docstrings (see the "Google decorators" page of the documentation).
 - Add the `deprecated.numpy` module: the `@deprecated`, `@versionadded` and `@versionchanged`
   decorators insert a `Deprecated`, `Version added` or `Version changed` section
   in [NumPy style](https://numpydoc.readthedocs.io/en/latest/format.html)